tests: integration: tap_hold: to separate modules
